setup algolia indexes for prod
