+++ src/app/[locale]/layout.tsx
@@ -23,6 +23,7 @@ export default async function LocaleLayout({
           <Navbar
             mainMenu={mainMenu}
             env={{
+              ALGOLIA_INDEX: process.env.ALGOLIA_INDEX!,
               ALGOLIA_APP_ID: process.env.ALGOLIA_APP_ID!,
               ALGOLIA_SEARCH_API_KEY: process.env.ALGOLIA_SEARCH_API_KEY!,
             }}
