+++ src/app/[locale]/posts/[category]/page.tsx
@@ -30,6 +30,7 @@ export default async function Page(props: LocaleProps) {
         categories={categories}
         topics={topics}
         env={{
+          ALGOLIA_INDEX: process.env.ALGOLIA_INDEX!,
           ALGOLIA_APP_ID: process.env.ALGOLIA_APP_ID!,
           ALGOLIA_SEARCH_API_KEY: process.env.ALGOLIA_SEARCH_API_KEY!,
         }}
