+++ src/app/[locale]/tutorials/page.tsx
@@ -6,6 +6,7 @@ export default function Page(props: Omit<Props, "env">) {
       <TutorialsPage
         {...props}
         env={{
+          ALGOLIA_INDEX: process.env.ALGOLIA_INDEX!,
           ALGOLIA_APP_ID: process.env.ALGOLIA_APP_ID!,
           ALGOLIA_SEARCH_API_KEY: process.env.ALGOLIA_SEARCH_API_KEY!,
         }}
