+++ src/blocks/dataBlocks/BlockCommunityEvents/BlockCommunityEvents.tsx
@@ -10,6 +10,7 @@ export async function BlockCommunityEvents(props: AutoProps): JSX.Element {
       <BlockCommunityEventsList
         {...props}
         env={{
+          ALGOLIA_INDEX: process.env.ALGOLIA_INDEX!,
           ALGOLIA_APP_ID: process.env.ALGOLIA_APP_ID!,
           ALGOLIA_SEARCH_API_KEY: process.env.ALGOLIA_SEARCH_API_KEY!,
         }}
